pyproject.toml
[tool.isort]
color_output = true
profile = "black"

[tool.poetry]
authors = ["adolf <adolf1321794021@gmail.com>"]
description = ""
name = "web3info"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
duckduckgo-search = "^6.2.13"
jinja2 = "^3.1.4"
markdownify = "^0.13.1"
openai = "^1.50.2"
pathvalidate = "^3.2.1"
pygments = "^2.18.0"
python = "^3.12"
python-dotenv = "^1.0.1"
seaborn = "^0.13.2"
serpapi = "^0.1.5"
trafilatura = "^1.12.2"
transformers = {extras = ["agent"], version = "^4.45.2"}

[tool.poetry.group.lint.dependencies]
ruff = "^0.6.8"
toml-sort = "^0.23.1"

[tool.ruff]
fix = true
line-length = 120

[tool.ruff.lint]
ignore = [
	"ANN", "D", "ERA001", "FIX", "INP001", "PGH", "PLR0913", "RUF012", "S", "T20", "TCH003", "TD"
]
select = ["ALL"]

[tool.ruff.per-file-ignores]
"docs/conf.py" = ["INP001"]
"test/*" = ["S101"]

[tool.tomlsort]
all = true
in_place = true
trailing_comma_inline_array = true
